Rename: Location 파일명 LocationWithCustomOverlay로 변경
